Enable binary import of latest(tip) versions
In case the maven repo that holds the suite to be imported does not
support the `latest` element in the XML metadata a fallback solution is
implemented that searches for the last known good artifact in the repo.
This allows us to use mx.Suite.import_suite for binary suites without
specifying the exact version (i.e. take the latest available)