Add LINKY_LAYOUT support
If the `LINKY_LAYOUT` environment variable is set, it is used as a glob
pattern. Files matching that pattern will be linked into the output
directory of layout distributions rather than copied.

When using "compile on save" in Eclipse or a similar workflow, having
```
LINKY_LAYOUT=*.jar
```
in `~/.mx/env` will allow automated update of the JARs in layout
distributions (such as the GraalVM).