Fix picking up version/url from dynamic imports in suites
Dynamic imports are specified on the command line or through the
environment. Usually, they are just resolved from the disk: the command
line or environment can't provide URLs or versions.
However some suite might provide versions for those imports by specifing
`"dynamic": True`. Such imports are only activated when the imported
suite is part of the command line/environment dynamic imports.

The "unversionned" dynamic imports from the command line/environment are
artificially added as imports of the primary suite with not version or
url. They are added last to give a chance to the "versioned" dynamic
imports to be resolved first.

However that is not enough when a dynamically imported suite is the one
that provides the version for an other dynamic import.
To solve this, we remove such "weak" unversionned dynamic imports from the
worklist when we append a versionned import.