[GR-29961] Improve compiler code coverage
The mx_gate currently records includes and excludes of components for
coverage reporting. The includes and excludes and not maintained
relative to each other and this results in conflicting options being
passed to JaCoCo leading to undefined behavior. This change makes the
includes and excludes mutually exclusive and allows a later include to
countermand an earlier exclude and vice versa.