Add support for merging layout dir distributions from multiple platforms
`mx archive-pd-layouts archive.tgz` can be used to create an archive
containing the output of all platform-dependent layout dir directories
currently built for the local platform.

This archive can be sent to another platform and can be restored with
`mx restore-pd-layouts archive.tgz`. This will add the files for the
first platform into mx's build directory.

This can be repeated for multiple platforms. Once all desired/necessary
platforms have been restored on one machines, a build can be started
with `mx --multi-platform-layout-directories=all build` which will cause
the platform-dependent layout directories to logically produce the union
of the files from all platforms.

For this to work, the platform-dependent layout directories must list
the platforms they support in their `platforms` attribute.

When using `--multi-platform-layout-directories=all`, it is a build
error if there are no files available on disk for a platform list by the
`platforms` attribute.

It is possible to limit the set of targetted platform by passing a
comma-separated list of platforms to
`--multi-platform-layout-directories=`.