Enable build support for JEP 493 enabled base JDKs
JEP 493-enabled JDKs won't have a 'jmods' directory by
default. For builds with option --no-jlinking the replacement
of base JDK-provided non-upgradable modules is not required,
nor desired. Allow for builds with no JMODs in that case by
checking for this option being in use and if so upgrade the
module path when compiling module-info.java files.